32-bit Cheshire

+++ sw/include/dif/uart.h
@@ -27,19 +27,19 @@
 #define UART_LINE_STATUS_THR_EMPTY_BIT 5
 #define UART_LINE_STATUS_TMIT_EMPTY_BIT 6
 
-void uart_init(void *uart_base, uint64_t freq, uint64_t baud);
+void uart_init(void *uart_base, uint32_t freq, uint32_t baud);
 
 int uart_read_ready(void *uart_base);
 
 void uart_write(void *uart_base, uint8_t byte);
 
-void uart_write_str(void *uart_base, void *src, uint64_t len);
+void uart_write_str(void *uart_base, void *src, uint32_t len);
 
 void uart_write_flush(void *uart_base);
 
 uint8_t uart_read(void *uart_base);
 
-void uart_read_str(void *uart_base, void *dst, uint64_t len);
+void uart_read_str(void *uart_base, void *dst, uint32_t len);
 
 // Default UART provides console
 void _putchar(char byte);
